+++ src/ontology/mod.rs
@@ -1,4 +1,4 @@
 mod error;
+mod github_ontology_registry;
 mod github_release_client;
-mod hpo_registry;
 mod traits;
